move files from catalog to inventory collection
