+++ imports/plugins/core/catalog/server/no-meteor/utils/publishProductToCatalog.js
@@ -39,6 +39,9 @@ export default async function publishProductToCatalog(product, context) {
     }
   };
 
+
+  console.log("---------- modifier", modifier);
+
   Schemas.Catalog.validate(modifier, { modifier: true });
 
   // Insert/update catalog document
