+++ imports/plugins/core/catalog/server/no-meteor/utils/publishProductToCatalog.test.js
@@ -314,6 +314,8 @@ test("expect true if a product is published to the catalog collection", async ()
   mockContext.collections.Products.findOne.mockReturnValue(Promise.resolve(updatedMockProduct));
   mockContext.collections.Catalog.updateOne.mockReturnValueOnce(Promise.resolve({ result: { ok: 1 } }));
   const spec = await publishProductToCatalog(mockProduct, mockContext);
+  console.log("------------- spec", spec);
+
   expect(spec).toBe(true);
 });
 
@@ -322,5 +324,6 @@ test("expect false if a product is not published to the catalog collection", asy
   mockContext.collections.Shops.findOne.mockReturnValueOnce(Promise.resolve(mockShop));
   mockContext.collections.Catalog.updateOne.mockReturnValueOnce(Promise.resolve({ result: { ok: 0 } }));
   const spec = await publishProductToCatalog(mockProduct, mockContext);
+  console.log("------------- spec2", spec);
   expect(spec).toBe(false);
 });
